smooooth progressbars

+++ Contractor/Contractor/Interfaces/IAppData.cs
@@ -10,7 +10,7 @@ namespace Contractor.Interfaces
     {
         public float Factor { get; set; }
 
-        public  int ProdSeconds { get; set; }
+        public  float ProdSeconds { get; set; }
         
         public  int MaxProductiveTime { get; set; }
