+++ Contractor/Contractor/Model/SaveDataModel.cs
@@ -10,7 +10,7 @@ namespace Contractor.Model
     public class SaveDataModel : IAppData
     {
         public float Factor { get; set; }
-        public int ProdSeconds { get; set; }
+        public float ProdSeconds { get; set; }
 
         public int FreeSeconds { get; set; }
         public int MaxProductiveTime { get; set; }
